+++ kittens/quick_access_terminal/main.go
@@ -67,6 +67,9 @@ func main(cmd *cli.Command, opts *Options, args []string) (rc int, err error) {
 	if runtime.GOOS != "darwin" {
 		argv = append(argv, fmt.Sprintf("--app-id=%s", conf.App_id))
 	}
+	if conf.Output_name != "" {
+		argv = append(argv, fmt.Sprintf("--output-name=%s", conf.Output_name))
+	}
 	argv = append(argv, fmt.Sprintf("--focus-policy=%s", conf.Focus_policy))
 	if conf.Start_as_hidden {
 		argv = append(argv, `--start-as-hidden`)